moving editpopup into UI folder
